--- src/app/Queries.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Queries() {
-  return (
-    <div>
-      <p>This is the Queries tab</p>
-    </div>
-  );
-}
-export default Queries;
