--- src/app/Store.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Store() {
-  return (
-    <div>
-      <p>This is the Store tab</p>
-    </div>
-  );
-}
-export default Store;
