--- src/custom.d.ts
@@ -1,5 +1,0 @@
-// config needed to render svg in Typescript React app
-// declare module '*.svg' {
-//   const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
-//   export default content;
-// }
